Update pom.xml dependencies and remove unused code
Updated the pom.xml dependencies to better match the project's requirements. Several lines of script in multiple .java files were removed to declutter the codebase. Also, the state of browser execution visibility was changed from headless to headful in PlaywrightTestBase.java, which allows users to view the browser's execution during the testing process.