Update log4j dependencies and add database source file
This commit modifies the pom.xml file to update the log4j dependencies by defining the version in the properties section and using that property in the dependencies list. Therefore, we need to update the version only at one place in future. Also, a new XML was added that defines a PostgreSQL database source. In addition, Playwright browser launch options were updated to launch in headless mode.