Expand SaveATrainE2ETests class documentation
This commit adds detailed descriptions of the class fields and methods within the SaveATrainE2ETests class. It helps to understand the purpose of each field and method used for performing end-to-end tests on the Save A Train application. This enhanced documentation will be extremely useful for any developer reviewing or working with the code.